Create cloudinary to handle image and generic for mapping data between model and entity
