--- be/src/main/java/lavie/skincare_booking/util/CustomValidationException.java
@@ -1,16 +1,0 @@
-package lavie.skincare_booking.util;
-
-import java.util.List;
-
-public class CustomValidationException extends RuntimeException {
-    private final List<String> errors;
-
-    public CustomValidationException(List<String> errors) {
-        super(String.join(", ", errors));
-        this.errors = errors;
-    }
-
-    public List<String> getErrors() {
-        return errors;
-    }
-}
