+++ be/src/main/java/lavie/skincare_booking/utils/CustomValidationException.java
@@ -1,0 +1,16 @@
+package lavie.skincare_booking.utils;
+
+import java.util.List;
+
+public class CustomValidationException extends RuntimeException {
+    private final List<String> errors;
+
+    public CustomValidationException(List<String> errors) {
+        super(String.join(", ", errors));
+        this.errors = errors;
+    }
+
+    public List<String> getErrors() {
+        return errors;
+    }
+}
