+++ be/src/main/java/lavie/skincare_booking/utils/ResponseUtil.java
@@ -1,0 +1,58 @@
+package lavie.skincare_booking.utils;
+
+import lavie.skincare_booking.model.MetaDataDTO;
+import lavie.skincare_booking.model.ResponseDTO;
+import org.springframework.beans.BeanUtils;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+public class ResponseUtil {
+    public static ResponseEntity<ResponseDTO> getObject(Object result, HttpStatus status, String response) {
+        return new ResponseEntity<>(
+                ResponseDTO.builder()
+                        .details(ExceptionUtils.getResponseString(response))
+                        .content(result)
+                        .statusCode(status.value())
+                        .build()
+                , status
+        );
+    }
+
+    public static ResponseEntity<?> getCollection(Object result, HttpStatus status, String response
+            , MetaDataDTO metaData) {
+        return new ResponseEntity<>(
+                ResponseDTO.builder()
+                        .statusCode(status.value())
+                        .details(ExceptionUtils.getResponseString(response))
+                        .content(result)
+                        .metaDataDTO(metaData)
+                        .build()
+                , status
+        );
+    }
+
+    public static ResponseEntity<?> error(String error, String message, HttpStatus status) {
+        return new ResponseEntity<>(
+                ResponseDTO.builder()
+                        .details(ExceptionUtils.getError(error))
+                        .statusCode(status.value())
+                        .build()
+                , status
+        );
+    }
+
+    //Generate response class
+    public static <S, T> T map(S source, Class<T> targetClass) {
+        if (source == null) {
+            return null;
+        }
+        try {
+            T target = targetClass.getDeclaredConstructor().newInstance();
+            BeanUtils.copyProperties(source, target);
+            return target;
+        } catch (Exception e) {
+            throw new RuntimeException("Error while mapping " + source.getClass().getSimpleName() +
+                    " to " + targetClass.getSimpleName(), e);
+        }
+    }
+}
